--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,5 @@
+module github.com/felixge/go-xxd
+
+go 1.16
+
+require github.com/ogier/pflag v0.0.1