--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,5 @@
 module github.com/felixge/go-xxd
 
-go 1.16
+go 1.18
 
 require github.com/ogier/pflag v0.0.1