Changes include:
- Idiomatic style
- Formatting
- Remove unused vars
- Use `go 1.18` in go.mod

Optimizations include:
- Eliminate unnecessary bound checks
  (-gcflags="-d=ssa/check_bce/debug=1"): Reduce from 54 to 46 checks.
- Inline function by reducing costs (eg. Convert op "range" to normal
  loops, etc.)
- Reuse vars instead of allocating them in every iteration, etc.

Benchmarks are taken on:
```
Go version: go1.16.3 linux/arm64
OS, Arch: linux (manjaro), arm64 (rpi)
```

Note:
```
Perflock is used for stable benchmarks and Benchstat to compute and
compare statistics about benchmarks.
```

Result:
$ go test -bench=. -benchmem -count=10

```
$ benchstat old.txt new.txt
name   old time/op    new time/op    delta
XXD-4    73.6ns ± 1%    72.8ns ± 0%  -1.13%  (p=0.000 n=10+10)

name   old alloc/op   new alloc/op   delta
XXD-4     0.00B          0.00B         ~     (all equal)

name   old allocs/op  new allocs/op  delta
XXD-4      0.00           0.00         ~     (all equal)
```