@shmsr 2022-04-14
Made this commit ~10 months back but forgot to open a pull-request and that's why the benchmark are taken against go1.16