@felixge 2022-04-14
Nice! I don't have time for in-depth review, but I skimmed the patch and it looks cool : ).

`-gcflags="-d=ssa/check_bce/debug=1"` I didn't know about this option, thanks! I've always been digging through the generated assembly, lol 🙈